renamed package data to datagen
